OrgPolicyPolicy: fix mapper and fuzzer
