+++ pkg/controller/direct/orgpolicy/policy_fuzzer.go
@@ -37,10 +37,12 @@ func orgPolicyPolicyFuzzer() fuzztesting.KRMFuzzer {
 	f.SpecFields.Insert(".dry_run_spec")
 	f.SpecFields.Insert(".etag")
 
-	f.StatusFields.Insert(".spec")
+	f.StatusFields.Insert(".spec.update_time")
+	f.StatusFields.Insert(".dry_run_spec.update_time")
 
 	f.UnimplementedFields.Insert(".name")
-	f.UnimplementedFields.Insert(".alternate")
-	f.UnimplementedFields.Insert(".spec.update_time") // Output only field
+	f.UnimplementedFields.Insert(".alternate")          // deprecated field
+	f.UnimplementedFields.Insert(".dry_run_spec.rules") // skip fuzzer on "oneof" boolean fields
+	f.UnimplementedFields.Insert(".spec.rules")         // skip fuzzer on "oneof" boolean fields
 	return f
 }
